+++ lib/version.c
@@ -26,7 +26,7 @@
 #include "urldata.h"
 #include "vtls/vtls.h"
 #include "http2.h"
-#include "ssh.h"
+#include "vssh/ssh.h"
 #include "quic.h"
 #include "curl_printf.h"
